+++ CHANGELOG.zh_CN.md
@@ -8,6 +8,7 @@
 - 表单新增 submitOnReset 控制是否在重置时重新发起请求
 - 表格新增`sortFn`支持自定义排序
 - 新增动画组件及示例
+- 新增懒加载/延时加载组件及示例
 
 ### ✨ Refactor
