+++ src/components/Container/index.ts
@@ -1,5 +1,5 @@
 export { default as ScrollContainer } from './src/ScrollContainer.vue';
 export { default as CollapseContainer } from './src/collapse/CollapseContainer.vue';
-export { default as LazyContainer } from './src/LazyContainer';
+export { default as LazyContainer } from './src/LazyContainer.vue';
 
 export * from './src/types.d';
