--- src/components/Container/src/LazyContainer.less
@@ -1,27 +1,0 @@
-.lazy-container-enter {
-  opacity: 0;
-}
-
-.lazy-container-enter-to {
-  opacity: 1;
-}
-
-.lazy-container-enter-from,
-.lazy-container-enter-active {
-  position: absolute;
-  top: 0;
-  width: 100%;
-  transition: opacity 0.3s 0.2s;
-}
-
-.lazy-container-leave {
-  opacity: 1;
-}
-
-.lazy-container-leave-to {
-  opacity: 0;
-}
-
-.lazy-container-leave-active {
-  transition: opacity 0.5s;
-}
